# stargate.client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Stargate.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/AstronautDuties": {
      "get": {
        "tags": [
          "AstronautDuties"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautDuty"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautDuty"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautDuty"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AstronautDuties"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautDuty"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautDuty"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautDuty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautDuty"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautDuty"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautDuty"
                }
              }
            }
          }
        }
      }
    },
    "/api/AstronautDuties/{id}": {
      "get": {
        "tags": [
          "AstronautDuties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautDuty"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautDuty"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautDuty"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AstronautDuties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautDuty"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautDuty"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautDuty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "AstronautDuties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AstronautDuties/GetAstronautDuties/{personId}": {
      "get": {
        "tags": [
          "AstronautDuties"
        ],
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautDuty"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautDuty"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautDuty"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/People": {
      "get": {
        "tags": [
          "People"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonViewModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonViewModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "People"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        }
      }
    },
    "/api/People/{id}": {
      "get": {
        "tags": [
          "People"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "People"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "People"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AstronautDuty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "personId": {
            "type": "integer",
            "format": "int32"
          },
          "rank": {
            "type": "string",
            "nullable": true
          },
          "dutyTitle": {
            "type": "string",
            "nullable": true
          },
          "dutyStartDate": {
            "type": "string",
            "format": "date-time"
          },
          "person": {
            "$ref": "#/components/schemas/Person"
          }
        },
        "additionalProperties": false
      },
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "astronaut": {
            "type": "boolean",
            "readOnly": true
          },
          "astronautDuties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AstronautDuty"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PersonViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "astronaut": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}